Lots of fix

+++ A3/PT2/A3_PT_2.c
@@ -6,7 +6,7 @@ int main() {
 
     /* Declare constants */
     int cm, finalFoot, finalInch;
-    double inch, foot;
+    int inch, foot;
 
     /* Ask client for value */
     printf("Please enter the length in cm: ");
